pyproject.toml
[project]
name = "certificate-service"
version = "0.1.0"
description = ""
authors = [
    {name = "myadegari",email = "myadegari37@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-pptx (>=1.0.2,<2.0.0)",
    "comtypes (>=1.4.11,<2.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "jdatetime (>=5.2.0,<6.0.0)",
    "pika (>=1.3.2,<2.0.0)",
    "pptxtopdf (>=0.0.2,<0.0.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
